+++ packages/handlersjs-http/lib/handlers/http-cors-request.handler.ts
@@ -1,24 +1,17 @@
 import { Observable, of } from 'rxjs';
-import { map } from 'rxjs//operators';
+import { map } from 'rxjs/operators';
 import { HttpHandler } from '../models/http-handler';
 import { HttpHandlerContext } from '../models/http-handler-context';
 import { HttpHandlerResponse } from '../models/http-handler-response';
 import { cleanHeaders } from '../util/clean-headers';
 
-/**
- * Abstract class representing Http CORS options.
- */
-export abstract class HttpCorsOptions {
-
-  constructor(
-    public origins?: string[],
-    public allowMethods?: string[],
-    public allowHeaders?: string[],
-    public exposeHeaders?: string[],
-    public credentials?: boolean,
-    public maxAge?: number,
-  ) { }
-
+export interface HttpCorsOptions {
+  origins?: string[];
+  allowMethods?: string[];
+  allowHeaders?: string[];
+  exposeHeaders?: string[];
+  credentials?: boolean;
+  maxAge?: number;
 }
 
 /**